Update rewrite.
Still not compatible with every site the original was compatible with, so not ready to be merged just yet. This rewrite will heavily reduce (but not altogether prevent) issues caused by filters interfering with sites they were never meant to touch. The downside is a much much much messier script and a more tedious filter making process. This could have been partially prevented by being a bit less rigid with this new approach, but I didn't want to do things halfway. Oh well.